Grain-based reminders: separate IReminderTable & IReminderTableGrain
