+++ src/Orleans.Core/SystemTargetInterfaces/IReminderTable.cs
@@ -46,9 +46,19 @@ namespace Orleans
     /// Reminder table interface for grain based implementation.
     /// </summary>
     [Unordered]
-    internal interface IReminderTableGrain : IGrainWithIntegerKey, IReminderTable
+    internal interface IReminderTableGrain : IGrainWithIntegerKey
     {
-        
+        Task<ReminderTableData> ReadRows(GrainReference key);
+
+        Task<ReminderTableData> ReadRows(uint begin, uint end);
+
+        Task<ReminderEntry> ReadRow(GrainReference grainRef, string reminderName);
+
+        Task<string> UpsertRow(ReminderEntry entry);
+
+        Task<bool> RemoveRow(GrainReference grainRef, string reminderName, string eTag);
+
+        Task TestOnlyClearTable();
     }
 
     public class ReminderTableData
